requirements.txt
Flask
imutils
tensorflow
opencv-python
Pillow
scikit-image